01042025 att

+++ src/main/java/application/repository/QuestaoRepository.java
@@ -3,5 +3,5 @@ package application.repository;
 import application.model.Questao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestaoRepository extends JpaRepository<Questao, Integer> {
+public interface QuestaoRepository extends JpaRepository<Questao, Long> {
 }
